+++ src/main/java/zenpad/runners/exec/JavaExec.java
@@ -5,8 +5,8 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.StandardOpenOption;
 
-import zenpad.misc.LogUtils;
 import zenpad.misc.factory.DialogFactory;
+import zenpad.misc.util.LogUtility;
 import zenpad.runners.TerminalLauncher;
 
 public class JavaExec implements CodeExec {
@@ -27,7 +27,7 @@ public class JavaExec implements CodeExec {
             compile.waitFor();
 
             if (compile.exitValue() != 0) {
-                LogUtils.logError(compile.getErrorStream());
+                LogUtility.logError(compile.getErrorStream());
                 DialogFactory.showError("Compilation failed. See compiler-error.log.", "Compilation error");
                 return false;
             }
